move v2 to separate module
